v7.4 open source release (#35)
* prepare for v7.4 release

* remove bazelversion restriction

* fix c++ tests

+++ src/Foundational/iwmisc/report_progress.cc
@@ -7,5 +7,5 @@
 
 #include "report_progress.h"
 
-template class Report_Progress_Template<unsigned int>;
-template int Report_Progress_Template<unsigned int>::initialise<Command_Line>(Command_Line&, char, int);
+template class Report_Progress_Template<uint64_t>;
+template int Report_Progress_Template<uint64_t>::initialise<Command_Line>(Command_Line&, char, int);
